# Conflicts:
#	deploy.sh

+++ src/main/java/com/berry/oss/dao/mapper/BucketInfoMapper.java
@@ -4,7 +4,6 @@ import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import com.berry.oss.dao.entity.BucketInfo;
 import com.berry.oss.module.dto.BucketStatisticsInfoDto;
 import com.berry.oss.module.vo.BucketInfoVo;
-import com.berry.oss.module.vo.BucketStatisticsInfoVo;
 import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
@@ -19,7 +18,7 @@ import java.util.List;
  */
 public interface BucketInfoMapper extends BaseMapper<BucketInfo> {
 
-    List<BucketInfoVo> listBucket(@Param("userId") Integer userId, @Param("name") String name);
+    List<BucketInfoVo> listBucket(@Param("userId") Long userId, @Param("name") String name);
 
-    List<BucketStatisticsInfoDto> getBucketUseInfo(@Param("userId") Integer userId);
+    List<BucketStatisticsInfoDto> getBucketUseInfo(@Param("userId") Long userId);
 }
